split DAUUtils to prevent LC classes to be loaded for non-DAU apps
