+++ flow-server/src/test/java/com/vaadin/flow/server/communication/ServerRpcHandlerTest.java
@@ -9,6 +9,7 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
+import org.mockito.ArgumentMatchers;
 import org.mockito.Mockito;
 
 import com.vaadin.flow.component.UI;
@@ -17,6 +18,7 @@ import com.vaadin.flow.component.internal.UIInternals;
 import com.vaadin.flow.function.DeploymentConfiguration;
 import com.vaadin.flow.internal.MessageDigestUtil;
 import com.vaadin.flow.internal.StateTree;
+import com.vaadin.flow.server.Constants;
 import com.vaadin.flow.server.VaadinRequest;
 import com.vaadin.flow.server.VaadinService;
 import com.vaadin.flow.server.VaadinSession;
@@ -43,6 +45,7 @@ public class ServerRpcHandlerTest {
 
     @Rule
     public ExpectedException thrown = ExpectedException.none();
+    private DeploymentConfiguration deploymentConfiguration;
 
     @Before
     public void setup() {
@@ -62,8 +65,7 @@ public class ServerRpcHandlerTest {
         Mockito.when(ui.getSession()).thenReturn(session);
         Mockito.when(ui.getCsrfToken()).thenReturn(csrfToken);
 
-        DeploymentConfiguration deploymentConfiguration = Mockito
-                .mock(DeploymentConfiguration.class);
+        deploymentConfiguration = Mockito.mock(DeploymentConfiguration.class);
         Mockito.when(service.getDeploymentConfiguration())
                 .thenReturn(deploymentConfiguration);
 
@@ -105,7 +107,9 @@ public class ServerRpcHandlerTest {
             @Override
             protected String getMessage(Reader reader) throws IOException {
                 return msg;
-            };
+            }
+
+            ;
         };
 
         ui = new UI();
@@ -125,7 +129,9 @@ public class ServerRpcHandlerTest {
             protected String getMessage(Reader reader) throws IOException {
                 return "{\"" + ApplicationConstants.CLIENT_TO_SERVER_ID
                         + "\":1}";
-            };
+            }
+
+            ;
         };
 
         ui = new UI();
@@ -137,6 +143,7 @@ public class ServerRpcHandlerTest {
     @Test(expected = DauEnforcementException.class)
     public void handleRpc_dauEnforcement_throws()
             throws InvalidUIDLSecurityKeyException, IOException {
+        enableDau();
         StringReader reader = new StringReader("{\"csrfToken\": \"" + csrfToken
                 + "\", \"rpc\":[{\"type\": \"event\", \"node\" : 1, \"event\": \"click\" }], \"syncId\": 0, \"clientId\":0}");
         ServerRpcHandler handler = new ServerRpcHandler();
@@ -152,6 +159,7 @@ public class ServerRpcHandlerTest {
     @Test
     public void handleRpc_dauEnforcement_pollEvent_doNoThrow()
             throws InvalidUIDLSecurityKeyException, IOException {
+        enableDau();
         StringReader reader = new StringReader("{\"csrfToken\": \"" + csrfToken
                 + "\", \"rpc\":[{\"type\": \"event\", \"node\" : 1, \"event\": \"ui-poll\" }], \"syncId\": 0, \"clientId\":0}");
         ServerRpcHandler handler = new ServerRpcHandler();
@@ -171,6 +179,7 @@ public class ServerRpcHandlerTest {
     @Test(expected = DauEnforcementException.class)
     public void handleRpc_dauEnforcement_pollEventMixedWithOtherEvents_throw()
             throws InvalidUIDLSecurityKeyException, IOException {
+        enableDau();
         StringReader reader = new StringReader("{\"csrfToken\": \"" + csrfToken
                 + "\", \"rpc\":[{\"type\": \"event\", \"node\" : 1, \"event\": \"ui-poll\" },{\"type\": \"event\", \"node\" : 1, \"event\": \"click\" }], \"syncId\": 0, \"clientId\":0}");
         ServerRpcHandler handler = new ServerRpcHandler();
@@ -186,6 +195,7 @@ public class ServerRpcHandlerTest {
     @Test(expected = ServerRpcHandler.ResynchronizationRequiredException.class)
     public void handleRpc_dauEnforcement_resynchronization_doNoThrow()
             throws InvalidUIDLSecurityKeyException, IOException {
+        enableDau();
         StringReader reader = new StringReader("{\"csrfToken\": \"" + csrfToken
                 + "\", \"rpc\":[{\"type\": \"event\", \"node\" : 1, \"event\": \"click\" }], \"resynchronize\": true, \"clientId\":0}");
         ServerRpcHandler handler = new ServerRpcHandler();
@@ -205,6 +215,7 @@ public class ServerRpcHandlerTest {
     @Test
     public void handleRpc_dauEnforcement_unloadBeacon_doNoThrow()
             throws InvalidUIDLSecurityKeyException, IOException {
+        enableDau();
         StringReader reader = new StringReader("{\"csrfToken\": \"" + csrfToken
                 + "\", \"rpc\":[{\"type\": \"event\", \"node\" : 1, \"event\": \"click\" }], \"UNLOAD\": true, \"clientId\":0}");
         ServerRpcHandler handler = new ServerRpcHandler();
@@ -224,6 +235,7 @@ public class ServerRpcHandlerTest {
     @Test
     public void handleRpc_dauEnforcement_returnChannelMessage_doNoThrow()
             throws InvalidUIDLSecurityKeyException, IOException {
+        enableDau();
         StringReader reader = new StringReader("{\"csrfToken\": \"" + csrfToken
                 + "\", \"rpc\":[{\"type\": \"channel\", \"node\" : 1, \"channel\": 0 }], \"syncId\": 0, \"clientId\":0}");
         ServerRpcHandler handler = new ServerRpcHandler();
@@ -243,6 +255,7 @@ public class ServerRpcHandlerTest {
     @Test(expected = DauEnforcementException.class)
     public void handleRpc_dauEnforcement_returnChannelMessageMixedWithOtherEvents_throw()
             throws InvalidUIDLSecurityKeyException, IOException {
+        enableDau();
         StringReader reader = new StringReader("{\"csrfToken\": \"" + csrfToken
                 + "\", \"rpc\":[{\"type\": \"channel\", \"node\" : 1, \"channel\": 0 },{\"type\": \"event\", \"node\" : 1, \"event\": \"click\" }], \"syncId\": 0, \"clientId\":0}");
         ServerRpcHandler handler = new ServerRpcHandler();
@@ -255,4 +268,11 @@ public class ServerRpcHandlerTest {
         handler.handleRpc(ui, reader, request);
     }
 
+    private void enableDau() {
+        Mockito.when(deploymentConfiguration.isProductionMode())
+                .thenReturn(true);
+        Mockito.when(deploymentConfiguration.getBooleanProperty(
+                ArgumentMatchers.eq(Constants.DAU_TOKEN),
+                ArgumentMatchers.anyBoolean())).thenReturn(true);
+    }
 }
