+++ flow-test-generic/src/main/java/com/vaadin/flow/testutil/ClassesSerializableTest.java
@@ -138,6 +138,9 @@ public abstract class ClassesSerializableTest extends ClassFinder {
                 "com\\.vaadin\\.flow\\.internal\\.nodefeature\\.NodeProperties",
                 "com\\.vaadin\\.flow\\.internal\\.AnnotationReader",
                 "com\\.vaadin\\.flow\\.server\\.StaticFileHandlerFactory",
+                "com\\.vaadin\\.flow\\.server\\.dau\\.DauEnforcementException",
+                "com\\.vaadin\\.flow\\.server\\.dau\\.FlowDauIntegration",
+                "com\\.vaadin\\.flow\\.server\\.dau\\.FlowDauIntegration\\$TrackingDetails",
                 "com\\.vaadin\\.flow\\.server\\.communication\\.ServerRpcHandler\\$LazyInvocationHandlers",
                 "com\\.vaadin\\.flow\\.server\\.VaadinServletRequest",
                 "com\\.vaadin\\.flow\\.server\\.VaadinServletResponse",
